Code cleanup. Moving path follower functionality out of ParkingBot and into new module.
Docstrings and removing unused functionality.
